Change extraction from denneys data.
Previously, I use as.double(attr(data$lat,"levels")), but Sophia found it was wrong.

I figure out that the "levels" is just a ordered possible value of all the latitude.

I correct this by using as.double(as.matrix(data.denneys$lat)). This gives us right order.

In that sense the previous plot is wrong, and probably the processed data results is wrong. I probably need change the writing up part for plot.